+++ CHANGELOG.md
@@ -19,7 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Windows port for Dialog DA14585 connected via serial port
 - AVRCP: introduced AVRCP_FEATURE_MASK_* as alternative to avrcp_controller_supported_feature_t and avrcp_target_supported_feature_t enums
 - AVDTP: renamed definition of supported features from AVDTP_SOURCE_SF_ and AVDTP_SINK_SF_ to AVDTP_SOURCE_FEATURE_MASK_ and AVDTP_SINK_FEATURE_MASK_ respectively. 
-
+- GAP: Provide gap_get_role to query master/slave role for HCI connection handle
+ 
 ### Changed
 - CVSD PLC: treat zero frames as good and allow to mark data as bad, e.g. if reported by controller as erroneous
 - btstack_audio: add set_volume() to sink and set_gain() to source interfaces